refactor: move the credits helpers into shared-frontend-utils
`apps/website` cannot import from the main app's `src/`: its tsconfig maps
only `@/* -> ./src/*` and the editor is not a package. So the Workshop
pages, which show a credit price per model run, had no way to reach
`CREDITS_PER_USD` and copied it instead — there is already a second `211`
in `apps/website/src/config/credits.ts` on the workshop branch.

The module was a good candidate to share as-is: pure conversion and
`Intl` formatting, no Vue, no app imports.

This follows the move `formatUtil` and `networkUtil` already made. The
file lands in the package, and `vite.config.mts` plus `tsconfig.json`
alias the old `@/base/credits/comfyCredits` path onto it, so all 17
importers are untouched — including three that `vi.mock` that exact path,
which keeps working for the same reason it does for `formatUtil` today.
