+++ packages/shared-frontend-utils/src/creditsUtil.ts
